+++ Tree Tracker/Navigation/MainFlowViewController.swift
@@ -1,16 +1,28 @@
 import UIKit
+import Photos
 
 final class MainFlowViewController: UITabBarController {
     override func viewDidLoad() {
         super.viewDidLoad()
 
-        let upload = UploadListFlowViewController()
+        let liveUpload = LiveUploadViewController(viewModel: LiveUploadViewModel(navigation: self))
+        let uploadQueue = UploadListFlowViewController()
+        let allTrees = NavigationViewController(rootViewController: TableListViewController(viewModel: TreesViewModel()))
 
-        let trees = NavigationViewController(rootViewController: TableListViewController(viewModel: TreesViewModel()))
+        viewControllers = [liveUpload, uploadQueue, allTrees]
 
-        viewControllers = [upload, trees]
+        liveUpload.tabBarItem = UITabBarItem(title: "", image: .cameraIcon, selectedImage: .cameraIcon)
+        uploadQueue.tabBarItem = UITabBarItem(title: "", image: .uploadIcon, selectedImage: .uploadIcon)
+        allTrees.tabBarItem = UITabBarItem(title: "", image: .listIcon, selectedImage: .listIcon)
 
-        upload.tabBarItem = UITabBarItem(title: "", image: .uploadIcon, selectedImage: .uploadIcon)
-        trees.tabBarItem = UITabBarItem(title: "", image: .listIcon, selectedImage: .listIcon)
+        selectedIndex = 1
+    }
+}
+
+extension MainFlowViewController: LiveUploadNavigating {
+    func triggerAskForDetailsAndStoreFlow(assets: [PHAsset], completion: @escaping (Bool) -> Void) {
+        DispatchQueue.main.async {
+            self.selectedViewController?.present(TreeDetailsFlowViewController(assets: assets, completion: completion), animated: true, completion: nil)
+        }
     }
 }